fix: Remove unnecessary non-PID 1 log output
Removed an `eprintln!` statement that logged when the process was
not running as PID 1. Observing the recent health-check logs, I find
this redundant, un-necessary and in some cases confusing.

The whole purpose of these logs is to ensure that the Devops engineer
is able to confirm that the application is being run as PID 1 and
something else is not configured as entry point in the docker
container. We already provide logs when it is currently run as PID
1. Right now, when we relaunch the process - it's is launched as a
child process and they log these extra message (pid1-rs: Process not
running as Pid 1) which makes all the logs confusing since we see
series of this message.